Significant progress made.
Functionality:
Basic leader elections (sans log checking) is working.

Servers in candidate mode are correctly interrupted if a leader is elected.

Servers can leave and come back freely on an individual basis. Have not implemented RPC  timeouts yet,
so if everybody leaves and comes back the cluster will halt since they are not resent.

Heartbeats and RequestVotes work, however once the log is imeplemented some cases will require log checking.

Not 100% threadsafe yet, although it seems to be for the current state of the code. Will need to add locks
to various timers and state variables.

Added some debugging output that can be switched on if necessary. It's usefullness is limited due to the nature
of multi-threaded code.